pyproject.toml
[tool.poetry]
name = "image-ops"
version = "0.1.0"
description = ""
authors = ["Aleksandr Evtushenko @uzkesh <uzkesh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
opencv-python-headless = "^4.9.0.80"
albumentations = "^1.3.1"
numpy = "^1.26.3"
pandas = "^2.2.0"
imagehash = "^4.3.1"
python-ulid = "^2.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.1.1"
flake8 = "^7.0.0"


[tool.poetry.group.cli.dependencies]
click = "^8.1.7"
tqdm = "^4.66.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
